# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treko API",
        "description": "Treko API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/donations": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Lista todas as doações do usuário autenticado",
                "operationId": "efd463fc838a5d0b3edcc89ecc8374ac",
                "responses": {
                    "200": {
                        "description": "Lista de doações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Donation"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Cria uma nova doação",
                "operationId": "a83d56537f533c01a1daf0abefdec913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "donated": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01 12:00:00"
                                    },
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doação criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Donation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O valor deve ser positivo."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/donations/{id}": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Obtém uma doação específica",
                "operationId": "getDonation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da doação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doação encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Donation"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Acesso não autorizado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doação não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Donations"
                ],
                "summary": "Atualiza uma doação",
                "operationId": "updateDonation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da doação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "donated": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.75
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-02 15:30:00"
                                    },
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doação atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Donation"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Donations"
                ],
                "summary": "Remove uma doação",
                "operationId": "ff34df7029ed4a4fc68e6a59674f9621",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da doação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Doação removida com sucesso"
                    },
                    "403": {
                        "description": "Acesso não autorizado"
                    },
                    "404": {
                        "description": "Doação não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Donation": {
                "required": [
                    "donated",
                    "date",
                    "user_id",
                    "campaign_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "donated": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "campaign_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}